style: entity 외 기타 변수명 수정
