+++ package.json
@@ -19,10 +19,12 @@
         "format": "prettier --write \"**/*.{ts,tsx,md,json}\""
     },
     "devDependencies": {
+        "@nestjs/testing": "^10.4.15",
         "@prettier/plugin-xml": "^3.1.1",
         "@trivago/prettier-plugin-sort-imports": "^4.3.0",
         "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-prettier": "^5.2.1",
+        "jest": "^29.7.0",
         "npm-run-all": "^4.1.5",
         "prettier": "^3.4.2"
     },
